Production of the maximal concentration
When multiple reactions are enabled, we produce the maximal concentration by exploiting the fact that reactions are sorted (increasing order).  We build the encoding starting from the inner-most reaction (with the lowest concentration level), and then build around that reaction encoding for the reactions with higher concentration levels. This allows us to avoid the disjunctions of possible resulting concentractions. E.g. (x==1 v x==3 v x==7), while we should have only x==7.